tcp implement

+++ src/xnet_app/xserver_datetime.c
@@ -20,7 +20,7 @@ xnet_err_t datetime_handler(xudp_t* udp, xipaddr_t* src_ip, uint16_t src_port, x
 	str_size = strftime((char*)tx_packet->data, TIME_STR_SIZE, "%A, %B, %d, %Y, %T-%z", timeinfo);
 
 	// 发送
-	truncate_packet(tx_packet, str_size);
+	truncate_packet(tx_packet, (uint16_t)str_size);
 	
 	return xudp_out(udp, src_ip, src_port, tx_packet);;
 }
